refactor: Step2 피드백 반영
