+++ README.md
@@ -1,6 +1,11 @@
 # android-signup
 
-## 기능 목록
+## Step2 기능 목록
 - [x] 타이틀 추가
 - [x] 입력 폼 추가
 - [x] 회원가입 버튼 추가
+
+## Step3 구현 목록
+- [x] Step2 피드백 반영
+- [] 
+- []
