+++ app/src/main/java/nextstep/signup/ui/theme/Dimens.kt
@@ -1,0 +1,13 @@
+package nextstep.signup.ui.theme
+
+import androidx.compose.ui.unit.dp
+
+object Dimens {
+    val LargePadding = 24.dp
+    val TopPadding = 112.dp
+    val StartPadding = 33.dp
+    val EndPadding = 33.dp
+    val ButtonHeight = 50.dp
+    val ButtonWidth = 280.dp
+    val ButtonRadius = 100.dp
+}
